Move applicant validation into it's own file
